pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-knowledge-assistant"
version = "0.1.0"
description = "Production-ready mini RAG system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic-settings>=2.4,<3.0",
  "python-dotenv>=1.0,<2.0",
  "httpx>=0.27,<0.28",
  "python-json-logger>=2.0,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "aiosqlite>=0.20,<1.0",
  "faiss-cpu>=1.8,<2.0",
  "sentence-transformers>=3.0,<4.0",
  "streamlit>=1.40,<2.0",
  "pypdf>=5.0,<6.0",
  "python-docx>=1.1,<2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.24,<1.0",
  "ruff>=0.6,<1.0",
  "mypy>=1.11,<2.0",
  "pre-commit>=3.8,<4.0"
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.core.logging"
disable_error_code = ["no-untyped-call"]
